refactor: rename Button component to PositionButton (suggested in PR)
- Renamed the Button component to PositionButton based on PR suggestion.
